feat: Make the number of words considered configurable
By default the code will look for the top 15 most interesting words in
the analysed text. *Interesting* in this case means words that deviate
the most from a probability of 0.5 of there being a match. With this
commit you can configured the number of words considered.

For example:

    Categoriser categoriser = new Categoriser()
        .category("business",
          new DirectorySampleSource("./data/samples/business"))
        .category("technology",
          new DirectorySampleSource("./data/samples/technology"))
        .category("sport",
          new DirectorySampleSource("./data/samples/sport"))
        .topWordsConsidered(200);

Will consider the top 200 words in each piece of analysed text.

If more words are looked at, then the match will become more accurate,
but more processing will be required.